Securize: Increase isolation by stripping away prototypes
When we assign our new `(secured-eval)` we've been allowing the `window`
property to bleed through as `this`. Additionally we have been setting
our forbidden symbols as the empty object `{}` which also exposes the
`Object` prototype.

In this patch we're replacing both of those in an attempt to further
limit the extent to which scripts can access global data.

The downside to this approach is that we've lost all prototypes that we
want, such as `Array.prototype.map`. The severity of this limitation is
so high that it's probably unmergable, but I'm hoping there might still
be a way to resolve that.